Jukebox Revelation
Ports all the old music, plus some new additions, of old persistent-bay. Updates the sprite to be rebased off of the old jukebox sprite, instead of being a TV. I converted the files that were compatible to OGG isntead of MIDI because OGG holds better sound quality in-game, and the majority of the game uses OGG as well.